pyproject.toml
[tool.poetry]
name = "kurs-5"
version = "0.1.0"
description = ""
authors = ["Ildar Gabdullin <ildar8803@gmail.com>"]
readme = "README.md"
packages = [{include = "kurs_5"}]

[tool.poetry.dependencies]
python = "^3.11"
psycopg2-binary = "^2.9.7"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
